feat: added venues test
