+++ uobtheatre/venues/test/factories.py
@@ -1,0 +1,14 @@
+import factory
+
+from uobtheatre.venues.models import (
+    Venue,
+)
+
+
+class VenueFactory(factory.django.DjangoModelFactory):
+
+    name = factory.Faker("sentence", nb_words=2)
+    internal_capacity = factory.Faker("pyint")
+
+    class Meta:
+        model = Venue
